Consolidate k8sclient: delete clientset.go, use NewClient() everywhere
- Add IsRunningInCluster() and getKubeConfig() to client.go
- NewClient() now supports both in-cluster and kubeconfig (local dev)
- Delete clientset.go (duplicate functionality)
- Update manager.go to use NewClient().GetClientset()
- Update tests to use NewClient().GetClientset()
